Bugfixes:
- add missing route on roles endpoint
- fix incorrect type on unit tests
- remove unnecessary directory
- fix integration tests

+++ tests/HRManagement.Staff.Tests/Features/Employees/UpdateEmployeeCommandHandlerShould.cs
@@ -40,7 +40,7 @@ public class UpdateEmployeeCommandHandlerShould
             .ReturnsAsync(true)
             .ReturnsAsync(true);
         mockRoleService
-            .Setup(x => x.CheckIfRoleExists(It.IsAny<byte>()))
+            .Setup(x => x.CheckIfRoleExists(It.IsAny<int>()))
             .ReturnsAsync(true);
 
         _mockEmployeeService
